Dump cache instead of rebuild on node upgrade
Node upgrades are so rare we don't save all that much rebuilding cache instead of re-fetching modules via npm.  Also node upgrades are liable to cause this rebuild to fail.  For example - node 0.8.x -> 0.10.x dropped node-waf.  Any cached modules that relied upon node-waf are going to cause the buildpack to fail before it even has an opportunity to fetch a newer version of the module dependent upon node-gyp instead of node-waf